Move remote control functionality to k9mail module
