+++ app/k9mail/src/main/java/com/fsck/k9/external/remotecontrol/RemoteControlService.java
@@ -1,23 +1,35 @@
-package com.fsck.k9.service;
+package com.fsck.k9.external.remotecontrol;
 
-import com.fsck.k9.Account;
-import com.fsck.k9.K9;
-import com.fsck.k9.preferences.Storage;
-import com.fsck.k9.preferences.StorageEditor;
-import com.fsck.k9.remotecontrol.K9RemoteControl;
-import com.fsck.k9.Preferences;
-import com.fsck.k9.core.R;
-import com.fsck.k9.Account.FolderMode;
-import com.fsck.k9.K9.BACKGROUND_OPS;
 
-import static com.fsck.k9.remotecontrol.K9RemoteControl.*;
+import java.util.List;
 
 import android.content.Context;
 import android.content.Intent;
-import timber.log.Timber;
 import android.widget.Toast;
 
-import java.util.List;
+import com.fsck.k9.Account;
+import com.fsck.k9.Account.FolderMode;
+import com.fsck.k9.K9;
+import com.fsck.k9.K9.BACKGROUND_OPS;
+import com.fsck.k9.Preferences;
+import com.fsck.k9.core.R;
+import com.fsck.k9.preferences.Storage;
+import com.fsck.k9.preferences.StorageEditor;
+import com.fsck.k9.service.BootReceiver;
+import com.fsck.k9.service.CoreService;
+import com.fsck.k9.service.MailService;
+import timber.log.Timber;
+
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_ACCOUNT_UUID;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_ALL_ACCOUNTS;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_BACKGROUND_OPERATIONS;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_NOTIFICATION_ENABLED;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_POLL_CLASSES;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_POLL_FREQUENCY;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_PUSH_CLASSES;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_RING_ENABLED;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_THEME;
+import static com.fsck.k9.external.remotecontrol.K9RemoteControl.K9_VIBRATE_ENABLED;
 
 public class RemoteControlService extends CoreService {
     private final static String RESCHEDULE_ACTION = "com.fsck.k9.service.RemoteControlService.RESCHEDULE_ACTION";
